moved modal and related files into folder
