+++ src/apps/dashboard/dashboard.tsx
@@ -2,9 +2,9 @@ import React, { FC, useCallback } from "react";
 import DashboardFees from "./dashboard-fees/dashboard-fees";
 import DashboardNotificationList from "./dashboard-notification-list/dashboard-notification-list";
 import { useText } from "../../core/utils/text";
-import StillInQueueModal from "./modal/still-in-queue-modal";
+import StillInQueueModal from "./modal/still-in-queue-modal/still-in-queue-modal";
 import { useModalButtonHandler } from "../../core/utils/modal";
-import StillInQueueModalContent from "./modal/still-in-queue-modal-content";
+import StillInQueueModalContent from "./modal/still-in-queue-modal/still-in-queue-modal-content";
 
 const DashBoard: FC = () => {
   const t = useText();
